+++ kryptone/kryptone-stubs/db/queries.pyi
@@ -49,5 +49,5 @@ class QuerySet:
     def __iter__(self) -> Generator[BaseRow]: ...
 
     def load_cache(self) -> None: ...
-    def exclude(self, **kwargs) -> QuerySet[BaseRow]: ...
-    def order_by(self, *fields) -> QuerySet[BaseRow]: ...
+    def exclude(self, **kwargs: str) -> QuerySet[BaseRow]: ...
+    def order_by(self, *fields: str) -> QuerySet[BaseRow]: ...
