+++ kryptone/kryptone-stubs/db/tables.pyi
@@ -31,7 +31,7 @@ class AbstractTable(metaclass=BaseTable):
 
 
 class Table(AbstractTable):
-    fields_map: OrderedDict = ...
+    fields_map: OrderedDict[str, Field] = ...
     name: str = ...
     query: Query = ...
